refactored code for better organizing followed https://github.com/Santos-Enoque/flutter-shopping-app
