+++ lib/constants/controllers.dart
@@ -1,0 +1,10 @@
+import 'package:the_asset_zone_web/controllers/nav_bar_controller.dart';
+import 'package:the_asset_zone_web/controllers/properties_controller.dart';
+import 'package:the_asset_zone_web/controllers/search_controller.dart';
+import 'package:the_asset_zone_web/controllers/single_page_property_controller.dart';
+
+NavBarController navBarController = NavBarController.instance;
+PropertyController propertyDetailsFirestore = PropertyController.instance;
+SearchController searchController = SearchController.instance;
+SinglePagePropertyController singlePagePropertyController =
+    SinglePagePropertyController();
