+++ lib/controllers/single_page_property_controller.dart
@@ -1,6 +1,7 @@
 import 'package:get/get.dart';
 
 class SinglePagePropertyController extends GetxController {
+  static SinglePagePropertyController instance = Get.find();
   final property = {
 
   }.obs;
