+++ lib/screens/home/components/home_screen_widgets.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:the_asset_zone_web/constants/constants.dart';
-import 'package:the_asset_zone_web/controllers/properties_state_controller.dart';
+import 'package:the_asset_zone_web/controllers/properties_controller.dart';
 import 'package:the_asset_zone_web/screens/single_property_page/single_page_property.dart';
 import '../../../widgets/helper_widgets.dart';
 import 'package:the_asset_zone_web/controllers/nav_bar_controller.dart';
