+++ lib/screens/single_property_page/components/single_page_property_middle_about_container.dart
@@ -1,7 +1,7 @@
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:the_asset_zone_web/controllers/properties_state_controller.dart';
+import 'package:the_asset_zone_web/controllers/properties_controller.dart';
 import 'package:the_asset_zone_web/models/property_detail_model.dart';
 
 class SinglePagePropertyMiddleAboutContainer extends StatelessWidget {
